Add QCT::InsertAs to address bug found by Greg
To wit: in Quiver sometimes we have no model trained for a named
chemistry (ex: XL-C2), so we want to use a model from some other
chemistry (ex: unknown) for reads of that chem.  We can only do this
if we break the tie between "chemistry that the model params were
trained on" and the "chemistry of the read we are to analyze".  We
do so via InsertAs(chemAlias, model).

[Tracking bug 25700]

[git-p4: depot-paths = "//depot/software/smrtanalysis/bioinformatics/lib/cpp/ConsensusCore/": change = 143326]