Potential fix for double-free issue found by @txje
ReadState's copy constructor and assignment operator were not
observing copy semantics.  This became apparent because the temporary
ReadState object constructed at MultiReadMutationScorer.cpp:281 was
being deleted---deleting the scorer within---but that scorer then also
belonged to a ReadState with the vector reads_.

A nicer solution would be possible in C++11---use emplace_back, delete
copy constructors all over the place... but we are going to retain
C++98 compliance for CC1.

A possible better fix would be to make ReadScore's fields be the
values instead of the pointers, but this would entail some more
refactoring... under consideration.

(problem found via valgrind...)